Fully working
Changed code to pass the whole Address object instead of just string for City and Province values from the FetchAddressIntentService back to resultReceivers (MainAcitivity and MapActivity).

Fixed code to display city name and first check for subLocaility and if null then get it from Locality and so on.

Add onClickListener to image for conditions to also show MapActivity